pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "nexo-enums"
version = "0.0.1"
description = "Enums package for Nexo"
authors = [
    { name = "Agra Bima Yuda", email = "agra@nexmedis.com" }
]
license = { text = "Proprietary" }
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "annotated-types>=0.7.0",
    "black>=25.1.0",
    "cfgv>=3.4.0",
    "click>=8.2.1",
    "colorama>=0.4.6",
    "distlib>=0.4.0",
    "filelock>=3.19.1",
    "identify>=2.6.13",
    "nexo-types>=0.0.1",
    "mypy_extensions>=1.1.0",
    "nodeenv>=1.9.1",
    "packaging>=25.0",
    "pathspec>=0.12.1",
    "platformdirs>=4.4.0",
    "pre_commit>=4.3.0",
    "pydantic>=2.12.2",
    "pydantic_core>=2.41.4",
    "PyYAML>=6.0.2",
    "typing-inspection>=0.4.2",
    "typing_extensions>=4.15.0",
    "virtualenv>=20.34.0",
]

[tool.setuptools]
packages = ["nexo.enums"]

[tool.setuptools.package-data]
"nexo.enums" = ["*.json", "*.yaml"]

[tool.setuptools.package-dir]
"nexo.enums" = "src"
